requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
requests==2.31.0
alembic==1.13.1
grpcio==1.60.0
grpcio-tools==1.60.0

